backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pandas==2.2.3
numpy==2.1.2
sqlmodel==0.0.21
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
passlib==1.7.4
PyJWT==2.9.0
orjson==3.10.7
reportlab==4.2.2
ortools==9.14.6206
pytest==8.3.3
httpx==0.27.2
starlette==0.38.5
